Tighten [Version] attribute to long-only + widen missed test class (#3733 CI fix)
CI on #4377 surfaced one missed [Version]-annotated property: the test
class MyAggregateWithDifferentVersionProperty had

    [Version]
    public int SpecialVersion { get; set; }

The original sed-based widening only matched the literal
"public int Version" property name; this one slipped through because
the property is called SpecialVersion. Widened to long.

Also tightened src/Marten/Schema/VersionAttribute.cs:
  * The Modify() branch that accepted int is removed — [Version] on a
    Marten 9 numeric-revisioning property must be long, since
    MetadataColumn<long> rejects an int member at registration time.
  * A dedicated diagnostic now fires when a user annotates an int
    property, naming the declaring type/member and pointing at the
    migration-guide section so the fix is obvious from the stack trace.
  * The catch-all error message dropped "int" from the supported types
    list to match the new contract.

Verified locally:
  * The previously-failing test
    EventSourcingTests.Aggregation.setting_version_number_on_aggregate.set_version_on_aggregate_with_explicit_Version_attribute
    passes.
  * Full EventSourcingTests on net9.0: 1274 passed, 2 skipped, 0 failed.

+++ src/EventSourcingTests/Aggregation/setting_version_number_on_aggregate.cs
@@ -132,7 +132,7 @@ public class setting_version_number_on_aggregate : OneOffConfigurationsContext
     public class MyAggregateWithDifferentVersionProperty
     {
         [Version]
-        public int SpecialVersion { get; set; }
+        public long SpecialVersion { get; set; }
 
 
         public Guid Id { get; set; }
